# src/api/docs/swagger.yaml
basePath: /
definitions:
  handler.APIResponse:
    properties:
      message:
        description: need fill only if error occurred
        type: string
    type: object
  handler.AuthAPIResponse:
    properties:
      access_token:
        type: string
      id:
        type: integer
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ListReceiptIngredientAPIResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/receipt.ReceiptIngredient'
        type: array
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ReceiptIngredientAPIResponse:
    properties:
      item:
        $ref: '#/definitions/receipt.ReceiptIngredient'
        type: object
      message:
        description: need fill only if error occurred
        type: string
    type: object
  ingredient.Ingredient:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  receipt.ReceiptIngredient:
    properties:
      created_at:
        type: string
      id:
        type: integer
      ingredient:
        $ref: '#/definitions/ingredient.Ingredient'
        type: object
      ingredient_id:
        type: integer
      quantity:
        type: string
      receipt_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  services.CreateReceiptIngredientRequest:
    properties:
      ingredient_id:
        minimum: 1
        type: integer
      quantity:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - ingredient_id
    - quantity
    type: object
...
  services.UpdateReceiptIngredientRequest:
    properties:
      quantity:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - quantity
    type: object
...
  user.SignInRequest:
    properties:
      password:
        description: (required)
        maxLength: 255
        minLength: 6
        type: string
      username:
        description: (required)
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.SignUpRequest:
    properties:
      password:
        description: (required)
        maxLength: 255
        minLength: 6
        type: string
      username:
        description: (required)
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
host: api.food.test
info:
  contact: {}
  description: This is Swagger docs for Food API
  license: {}
  title: Food API
  version: "1.0"
paths:
  /user/signIn:
    post:
      consumes:
      - application/json
      description: get a user by params
      parameters:
      - description: User params
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/user.SignInRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      summary: Get a user from the DB
      tags:
      - auth
  /user/signUp:
    post:
      consumes:
      - application/json
      description: create new user by params
      parameters:
      - description: User params
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/user.SignUpRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      summary: Add a new user to the DB
      tags:
      - auth
...
  /v1/receipts/{id}/ingredients:
    get:
      description: find receipt ingredients by params
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListReceiptIngredientAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get receipt ingredients
      tags:
      - receipts
    post:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.CreateReceiptIngredientRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptIngredientAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create receipt ingredient
      tags:
      - receipts
  /v1/receipts/{id}/ingredients/{ingredient_id}:
    delete:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: Receipt ingredient id
        in: path
        name: ingredient_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete receipt ingredient
      tags:
      - receipts
    put:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: Receipt ingredient id
        in: path
        name: ingredient_id
        required: true
        type: integer
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.UpdateReceiptIngredientRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptIngredientAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update receipt
      tags:
      - receipts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"